% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-basics.R
\name{annotate_icinac}
\alias{annotate_icinac}
\title{Annotate a table returned by \code{\link{read_icinac}}}
\usage{
annotate_icinac(tab.data, tab.anno, chan.name = "chan")
}
\arguments{
\item{tab.data}{data frame object returned by \code{\link{read_icinac}}}

\item{tab.anno}{data frame object containing the annotation}

\item{chan.name}{name of the column that contains the Channel information}
}
\value{
Returns tab.data object appended with the columns from the tab.anno
object.
}
\description{
This function is a wrapper, using the \code{\link{dplyr::left_join}} function
to append the data table with supplementary information, so that it can be used
for statistical analysis.
}
