% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{find_extremes}
\alias{find_extremes}
\alias{peaks}
\alias{pits}
\title{Finding extremes in a data set}
\usage{
find_extremes(x, column = NULL, pits = TRUE)

peaks(x)

pits(x)
}
\arguments{
\item{x}{A data frame or a vector of values}

\item{column}{If x is a data frame, the name of the column needs to be given as a character string.}

\item{pits}{Boolean: Should the function filter pits or peaks?}
}
\value{
A vector of Boolean values with the length of x, returning true if the respective point is an extreme
}
\description{
Finding extremes in a data set
}
\section{Functions}{
\itemize{
\item \code{peaks}: Wrapper for finding peaks

\item \code{pits}: Wrapper for finding pits
}}

\examples{
find_extremes(InsectSprays,"count") #finds minima
# or without column argument
find_extremes(InsectSprays$count,pits=FALSE) #finds maxima
# wrappers
peaks(InsectSprays$count)
pits(InsectSprays$count)
# example using functions from dplyr package
dplyr::filter(InsectSprays,peaks(count))
}
