% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-basics.R
\name{import_icinac}
\alias{import_icinac}
\title{A wrapper to import and annotate iCinac data with one function}
\usage{
import_icinac(path.data, path.anno, chan.name = "chan")
}
\arguments{
\item{path.data}{File containing iCinac data to be imported}

\item{path.data}{File containing matchin annotations}
}
\value{
A data frame
}
