% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., file, cols = 1, byrow = TRUE)
}
\arguments{
\item{...}{ggplot objects to be combined in one image.}

\item{cols}{Desired number of columns in the final image.}
}
\value{
A combined image of multiple ggplot objects on a single page.
}
