% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-basics.R
\name{read_icinac}
\alias{read_icinac}
\title{Import data frames from experiments with the iCinac(R) platform}
\usage{
read_icinac(path)
}
\arguments{
\item{path}{File containing the exported data.}
}
\value{
A data frame object. It has a single row for each exported variable. Furthermore a
column informing about channels is added.
}
\description{
This is the core function of this package. It converts the output of the iCinac software
into a tidy data frame object in R. The exported file must contain meta-data. It can either
be the data from a single probe or bundled data from multiple .ici files. Ideal, it is the
bundled data from a single experiment.
}
